Fixed residual keypresses carrying over to gameplay
If a key was pressed on a menu screen and then a world was opened, then the key
would be read as if it were pressed while the world was loaded. When the user
created a world with the letter 'e' in its name, the inventory would open on
world creation. Now, keyboard inputs are updated on key release as well as key
press, thereby ficing the issue.